refactor: use current_user_id from JWT auth in controllers
- AccountController: Get user_id from conn.assigns.current_user_id instead of params
- BankConnectionController: Use authenticated user_id from assigns
- AssetController: Remove @default_user_id workaround, use current_user_id
- Remove all TODO comments about getting user_id from auth

+++ backend/lib/wealth_backend_web/controllers/account_controller.ex
@@ -6,19 +6,15 @@ defmodule WealthBackendWeb.AccountController do
 
   action_fallback WealthBackendWeb.FallbackController
 
-  # TODO: Get user_id from authenticated session/JWT token
-  # For now, use a default test user_id
-  @default_user_id 1
-
-  def index(conn, params) do
-    user_id = Map.get(params, "user_id", @default_user_id)
+  def index(conn, _params) do
+    user_id = conn.assigns.current_user_id
     accounts = Accounts.list_accounts(user_id)
     render(conn, :index, accounts: accounts)
   end
 
-  def create(conn, %{"account" => account_params} = params) do
-    # Add default user_id if not provided
-    account_params = Map.put_new(account_params, "user_id", Map.get(params, "user_id", @default_user_id))
+  def create(conn, %{"account" => account_params}) do
+    user_id = conn.assigns.current_user_id
+    account_params = Map.put(account_params, "user_id", user_id)
     
     with {:ok, %Account{} = account} <- Accounts.create_account(account_params) do
       conn
@@ -29,23 +25,48 @@ defmodule WealthBackendWeb.AccountController do
   end
 
   def show(conn, %{"id" => id}) do
+    user_id = conn.assigns.current_user_id
     account = Accounts.get_account!(id)
-    render(conn, :show, account: account)
+    
+    # Ensure account belongs to authenticated user
+    if account.user_id == user_id do
+      render(conn, :show, account: account)
+    else
+      conn
+      |> put_status(:forbidden)
+      |> json(%{error: "Access denied"})
+    end
   end
 
   def update(conn, %{"id" => id, "account" => account_params}) do
+    user_id = conn.assigns.current_user_id
     account = Accounts.get_account!(id)
 
-    with {:ok, %Account{} = account} <- Accounts.update_account(account, account_params) do
-      render(conn, :show, account: account)
+    # Ensure account belongs to authenticated user
+    if account.user_id == user_id do
+      with {:ok, %Account{} = account} <- Accounts.update_account(account, account_params) do
+        render(conn, :show, account: account)
+      end
+    else
+      conn
+      |> put_status(:forbidden)
+      |> json(%{error: "Access denied"})
     end
   end
 
   def delete(conn, %{"id" => id}) do
+    user_id = conn.assigns.current_user_id
     account = Accounts.get_account!(id)
 
-    with {:ok, %Account{}} <- Accounts.delete_account(account) do
-      send_resp(conn, :no_content, "")
+    # Ensure account belongs to authenticated user
+    if account.user_id == user_id do
+      with {:ok, %Account{}} <- Accounts.delete_account(account) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      conn
+      |> put_status(:forbidden)
+      |> json(%{error: "Access denied"})
     end
   end
 end
